Moved files into the proper namespace.
