+++ lib/possessive/inflector/inflections.rb
