+++ lib/possessive/inflector/methods.rb
